requirement.txt
astor=0.8.1
beautifulsoup4=4.11.1
blas=1.0
bottleneck=1.3.5
brotli=1.0.9
brotli-bin=1.0.9
brotlipy=0.7.0
bzip2=1.0.8
ca-certificates=2024.12.14
certifi=2024.8.30
cffi=1.15.1
charset-normalizer=2.0.4
click=8.0.4
colorama=0.4.6
cryptography=38.0.2
cuda-cccl=12.4.127
cuda-cudart=11.7.99
cuda-cudart-dev=11.7.99
cuda-cupti=11.7.101
cuda-libraries=11.7.1
cuda-libraries-dev=11.7.1
cuda-nvrtc=11.7.99
cuda-nvrtc-dev=11.7.99
cuda-nvtx=11.7.91
cuda-runtime=11.7.1
cycler=0.11.0
dgl=1.1.2.cu117
et_xmlfile=1.1.0
fftw=3.3.9
flask=2.2.2
flit-core=3.6.0
fonttools=4.25.0
freetype=2.12.1
gettext=0.19.8.1
giflib=5.2.1
glib=2.78.4
glib-tools=2.78.4
gst-plugins-base=1.18.5
gstreamer=1.18.5
icc_rt=2022.1.0
icu=58.2
idna=3.4
importlib-metadata=4.11.3
intel-openmp=2021.4.0
itsdangerous=2.0.1
jinja2=3.1.2
joblib=1.1.1
jpeg=9e
kiwisolver=1.4.4
krb5=1.21.3
lerc=3.0
libbrotlicommon=1.0.9
libbrotlidec=1.0.9
libbrotlienc=1.0.9
libclang=12.0.1
libclang13=14.0.6
libcublas=11.10.3.66
libcublas-dev=11.10.3.66
libcufft=10.7.2.124
libcufft-dev=10.7.2.124
libcurand=10.3.5.147
libcurand-dev=10.3.5.147
libcusolver=11.4.0.1
libcusolver-dev=11.4.0.1
libcusparse=11.7.4.91
libcusparse-dev=11.7.4.91
libdeflate=1.17
libffi=3.4.4
libglib=2.78.4
libiconv=1.16
libnpp=11.7.4.75
libnpp-dev=11.7.4.75
libnvjpeg=11.8.0.2
libnvjpeg-dev=11.8.0.2
libogg=1.3.5
libpng=1.6.39
libpq=16.6
libprotobuf=3.6.0
libtiff=4.5.1
libuv=1.44.2
libvorbis=1.3.7
libwebp=1.2.4
libwebp-base=1.2.4
libxml2=2.9.14
libxslt=1.1.35
lxml=4.9.1
lz4-c=1.9.4
markupsafe=2.1.1
matplotlib=3.5.3
matplotlib-base=3.5.3
mkl=2021.4.0
mkl-service=2.4.0
mkl_fft=1.3.1
mkl_random=1.2.2
munkres=1.1.4
mysql-connector-c=6.1.11
mysql-connector-python=8.0.18
networkx=2.6.3
numexpr=2.8.4
numpy=1.21.5
numpy-base=1.21.5
openpyxl=3.0.10
openssl=3.4.0
packaging=22.0
pandas=1.3.5
pcre2=10.42
pillow=9.4.0
pip=22.3.1
ply=3.11
protobuf=3.6.0
psutil=5.9.0
pycparser=2.21
pyopenssl=23.0.0
pyparsing=3.0.9
pyqt=5.15.7
pyqt5-sip=12.11.0
pysocks=1.7.1
python=3.7.1
python-dateutil=2.8.2
python_abi=3.7
pytorch=1.13.1
pytorch-cuda=11.7
pytorch-mutex=1.0
pytz=2022.7
qt-main=5.15.2
qt-webengine=5.15.9
qtwebkit=5.212
requests=2.28.1
scikit-learn=1.0.2
scipy=1.7.3
setuptools=65.6.3
sip=6.6.2
six=1.16.0
soupsieve=2.3.2.post1
sqlite=3.45.3
threadpoolctl=2.2.0
tk=8.6.14
toml=0.10.2
torchaudio=0.13.1
torchvision=0.14.1
tornado=6.2
tqdm=4.64.1
typing_extensions=4.4.0
ucrt=10.0.22621.0
urllib3=1.26.14
vc=14.2
vc14_runtime=14.42.34433
vs2015_runtime=14.42.34433
werkzeug=2.2.2
wheel=0.38.4
win_inet_pton=1.1.0
wincertstore=0.2
xz=5.4.6
zipp=3.11.0
zlib=1.2.13
zstd=1.5.5
